made compute-alphas-MCMC fast to be default
