--- src/compute-alphas-MCMC-fast.c
@@ -1,127 +1,0 @@
-#include "combin.h"
-#include "tinygraph.h"
-#include "blant.h"
-#include "misc.h"
-#include <stdio.h>
-
-Gint_type _canonList[MAX_CANONICALS], _alphaList[MAX_CANONICALS];
-char _canonNumEdges[MAX_CANONICALS];
-
-Gint_type CountPath(TINY_GRAPH* g, TSET seen, TSET candidates, int k, int u, int v) {
-    seen = TSetUnion(seen, candidates);
-
-    Gint_type outbound_u[k], outbound_v[k], outboundSize_u = 0, outboundSize_v = 0;
-    if (TSetCardinality(seen) == k) return 1;
-
-    //Fill outbound for seen set
-    int i, j;
-
-    for (j = 0; j < k; j++) {
-        if(TSetIn(seen, j)) continue;
-
-        if(TinyGraphAreConnected(g, u, j)) {
-            outbound_u[outboundSize_u++] = j;
-        }
-
-        if(TinyGraphAreConnected(g, v, j)) {
-            outbound_v[outboundSize_v++] = j;
-        }
-    }
-
-
-    Gint_type calculatedValue_u[k], calculatedValue_v[k];
-    for (i = 0; i < k; i++) calculatedValue_u[i] = 0, calculatedValue_v[i] = 0;
-
-    //Recurse through neighbors in outbound
-    Gint_type total = 0;
-
-    for (i = 0; i < outboundSize_u; i++) {
-        if (!calculatedValue_u[outbound_u[i]]) {
-            TSetEmpty(candidates);
-            TSetAdd(candidates, outbound_u[i]);
-            calculatedValue_u[outbound_u[i]] = CountPath(g, seen, candidates, k, u, outbound_u[i]);
-        }
-        total += calculatedValue_u[outbound_u[i]];
-    }
-
-    for(i = 0; i < outboundSize_v; i++) {
-        if (!calculatedValue_v[outbound_v[i]]) {
-            TSetEmpty(candidates);
-            TSetAdd(candidates, outbound_v[i]);
-            calculatedValue_v[outbound_v[i]] = CountPath(g, seen, candidates, k, outbound_v[i], v);
-        }
-        total += calculatedValue_v[outbound_v[i]];
-    }
-    return total;
-}
-
-Gint_type ComputeAlphaNode(TINY_GRAPH* g, int k) {
-    Gint_type total = 0, i, j;
-    TSET seen = 0, candidates;
-    for (i = 0; i < k; i++) {
-        for (j = i+1; j < k; j++) {
-            if (!TinyGraphAreConnected(g, i, j))
-                continue;
-
-            TSetEmpty(candidates);
-            TSetAdd(candidates, i);
-            TSetAdd(candidates, j);
-
-            total += CountPath(g, seen, candidates, k, i, j);
-            TSetEmpty(candidates);
-        }
-    }
-    return total;
-}
-
-SET* _connectedCanonicals;
-int main(int argc, char* argv[]) {
-    if (argc != 2 && argc != 3 && argc != 4) {
-        fprintf(stderr, "USAGE: %s k\nOr,  %s k ID\nOr,  %s k start end [inclusive]\n", argv[0], argv[0], argv[0]);
-        exit(-1);
-    }
-    int k = atoi(argv[1]);
-    char BUF[BUFSIZ];
-#if SELF_LOOPS
-    TINY_GRAPH *g = TinyGraphSelfAlloc(k);
-#else
-    TINY_GRAPH *g = TinyGraphAlloc(k);
-#endif
-    _connectedCanonicals = canonListPopulate(BUF, _canonList, k, _canonNumEdges);
-    Gordinal_type numCanon = _connectedCanonicals->maxElem;
-
-	int start, end;
-	if(argc == 2) {
-		start = 0, end = numCanon - 1;
-	} else if (argc == 3) {
-		start = atoi(argv[2]), end = atoi(argv[2]);
-		if(start < 0 || start >= numCanon) {
-			fprintf(stderr, "Invalid ID\n");
-			exit(-1);
-		}
-	} else {
-		start = atoi(argv[2]), end = atoi(argv[3]);
-		if(start < 0 || start >= numCanon || end < 0 || end >= numCanon || start > end) {
-			fprintf(stderr, "Invalid range\n");
-			exit(-1);
-		}
-	}
-
-    int i;
-    for (i = start; i <= end; i++) {
-	Int2TinyGraph(g, _canonList[i]);
-    if (!SetIn(_connectedCanonicals, i)) _alphaList[i] = 0;
-    else _alphaList[i] = ComputeAlphaNode(g, k)/2;
-    }
-
-    printf(GORDINAL_FMT "\n", numCanon);
-    for (i = 0; i < numCanon; i++) {
-	printf(GINT_FMT " ", _alphaList[i]);
-    }
-    printf("\n");
-
-    TinyGraphFree(g);
-    SetFree(_connectedCanonicals);
-
-    return 0;
-}
